Reformatting code

+++ tests/preprocessing/test_batch_copy.py
@@ -13,6 +13,7 @@
 #     limitations under the License.
 
 import os
+
 # (C) Copyright IBM Corporation 2017, 2018, 2019
 # U.S. Government Users Restricted Rights:  Use, duplication or disclosure restricted
 # by GSA ADP Schedule Contract with IBM Corp.
@@ -34,7 +35,6 @@ class TestReshaper(TestCase):
         pass
 
     def test_copy(self):
-
         batch_sizer = MemorySizeEval(memory_tol_percent=0.1)
 
         Nx = int(3)
@@ -58,14 +58,12 @@ class TestReshaper(TestCase):
         for data_interval in data_interval_list:
             for batch_size in batch_size_list:
                 with make_temp_directory() as tmp_dir:
-
                     test_data = os.path.join(tmp_dir, f"data.h5")
                     with h5py.File(test_data, "w") as fp:
-
                         dataset = fp.create_dataset(
                             "data",
                             shape=(Nt, Nx, Ny, 1),
-                            dtype=[("T", np.float), ("X", np.float), ("Y", np.float)],
+                            dtype=[("T", float), ("X", float), ("Y", float)],
                         )
 
                         dataset[...] = Z_array
